+++ front/src/app/layout.tsx
@@ -2,6 +2,8 @@ import "@/app/globals.css";
 import Footer from "@/components/Footer";
 import Navbar from "@/components/Navbar";
 import { ContextProvider } from "@/context/GeneralContext";
+import Chatbot from "@/components/Chatbot";
+
 
 export default function RootLayout({
   children,
@@ -14,6 +16,7 @@ export default function RootLayout({
         <ContextProvider>
           <Navbar />
           <main>{children}</main>
+          <Chatbot/>
           <Footer />
         </ContextProvider>
       </body>
